Renamed low.py as interface.py and moved file into model dir
