+++ dpscope/model/interface.py
